requirements.txt
streamlit==1.31.0
firebase-admin==6.2.0
pillow==10.1.0
opencv-python-headless==4.8.1.78
python-dotenv==1.0.0
requests==2.31.0 